.GuardianFlask/requirements.txt
flask==3.0.3
flask-sqlalchemy==3.1.1
flask-login==0.6.3
flask-mail==0.10.0
pymysql==1.1.1
python-dotenv==1.0.1
itsdangerous==2.2.0
werkzeug==3.0.3
cryptography==42.0.8